+++ sdk/src/client/node_manager/mod.rs
@@ -28,7 +28,7 @@ use crate::{
 
 // The node manager takes care of selecting node(s) for requests until a result is returned or if quorum is enabled it
 // will send the requests for some endpoints to multiple nodes and compares the results.
-pub(crate) struct NodeManager {
+pub struct NodeManager {
     pub(crate) primary_node: Option<Node>,
     primary_pow_node: Option<Node>,
     pub(crate) nodes: HashSet<Node>,