+++ sdk/src/client/node_manager/syncing.rs
@@ -4,38 +4,39 @@
 #[cfg(not(target_family = "wasm"))]
 use {
     crate::types::{api::core::response::InfoResponse, block::protocol::ProtocolParameters},
-    std::collections::HashMap,
-    std::sync::Arc,
     std::{collections::HashSet, time::Duration},
     tokio::time::sleep,
 };
 
-use super::Node;
-use crate::client::{ClientInner, Error, Result};
+use super::{Node, NodeManager};
+use crate::client::{Client, ClientInner, Error, Result};
 
 impl ClientInner {
     /// Get a node candidate from the healthy node pool.
-    pub fn get_node(&self) -> Result<Node> {
-        if let Some(primary_node) = &self.node_manager.primary_node {
+    pub async fn get_node(&self) -> Result<Node> {
+        if let Some(primary_node) = &self.node_manager.read().await.primary_node {
             return Ok(primary_node.clone());
         }
 
-        let pool = self.node_manager.nodes.clone();
+        let pool = self.node_manager.read().await.nodes.clone();
 
         pool.into_iter().next().ok_or(Error::HealthyNodePoolEmpty)
     }
 
     /// returns the unhealthy nodes.
     #[cfg(not(target_family = "wasm"))]
-    pub fn unhealthy_nodes(&self) -> HashSet<&Node> {
-        self.node_manager
+    pub async fn unhealthy_nodes(&self) -> HashSet<Node> {
+        let node_manager = self.node_manager.read().await;
+
+        node_manager
             .healthy_nodes
             .read()
             .map_or(HashSet::new(), |healthy_nodes| {
-                self.node_manager
+                node_manager
                     .nodes
                     .iter()
                     .filter(|node| !healthy_nodes.contains_key(node))
+                    .cloned()
                     .collect()
             })
     }
@@ -45,7 +46,7 @@ impl ClientInner {
 impl ClientInner {
     /// Sync the node lists per node_sync_interval milliseconds
     pub(crate) async fn start_sync_process(
-        self: Arc<Self>,
+        &self,
         nodes: HashSet<Node>,
         node_sync_interval: Duration,
         ignore_node_health: bool,
@@ -61,6 +62,8 @@ impl ClientInner {
     }
 
     pub(crate) async fn sync_nodes(&self, nodes: &HashSet<Node>, ignore_node_health: bool) -> Result<()> {
+        use std::collections::HashMap;
+
         log::debug!("sync_nodes");
         let mut healthy_nodes = HashMap::new();
         let mut network_nodes: HashMap<String, Vec<(InfoResponse, Node)>> = HashMap::new();
@@ -99,10 +102,7 @@ impl ClientInner {
 
         if let Some(nodes) = network_nodes.get(most_nodes.0) {
             if let Some((info, _node_url)) = nodes.first() {
-                let mut network_info = self
-                    .network_info
-                    .write()
-                    .map_err(|_| crate::client::Error::PoisonError)?;
+                let mut network_info = self.network_info.write().await;
 
                 network_info.latest_milestone_timestamp = info.status.latest_milestone.timestamp;
                 network_info.protocol_parameters = ProtocolParameters::try_from(info.protocol.clone())?;
@@ -116,10 +116,46 @@ impl ClientInner {
         // Update the sync list.
         *self
             .node_manager
+            .read()
+            .await
             .healthy_nodes
             .write()
             .map_err(|_| crate::client::Error::PoisonError)? = healthy_nodes;
 
+        Ok(())
+    }
+}
+
+impl Client {
+    #[cfg(not(target_family = "wasm"))]
+    pub async fn update_node_manager(&self, node_manager: NodeManager) -> crate::wallet::Result<()> {
+        let node_sync_interval = node_manager.node_sync_interval;
+        let ignore_node_health = node_manager.ignore_node_health;
+        let nodes = node_manager
+            .primary_node
+            .iter()
+            .chain(node_manager.nodes.iter())
+            .cloned()
+            .collect();
+
+        *self.node_manager.write().await = node_manager;
+
+        self.sync_nodes(&nodes, ignore_node_health).await?;
+        let client = self.clone();
+
+        let sync_handle = tokio::spawn(async move {
+            client
+                .start_sync_process(nodes, node_sync_interval, ignore_node_health)
+                .await
+        });
+
+        *self._sync_handle.write().await = crate::client::SyncHandle(Some(sync_handle));
+        Ok(())
+    }
+
+    #[cfg(target_family = "wasm")]
+    pub async fn update_node_manager(&self, node_manager: NodeManager) -> crate::wallet::Result<()> {
+        *self.node_manager.write().await = node_manager;
         Ok(())
     }
 }